Added set_xx methods for single value fields
Fields:
alert_notes_present
minimum_severity
threat_notes_present

Tested manually, verified post request was gerated correctly.
Pending other PRs to modify unit test files
    alert_list = cb.select(Alert).set_minimum_severity(3)
    alert_list = cb.select(Alert).set_threat_notes_present(False)
    alert_list = cb.select(Alert).set_alert_notes_present(False)
